Gracefully handle array values of latest_receipt_info JSON
Despite the Apple documentation's claim that this is "only returned
for iOS 6 style transaction receipts", it appears to be returned for
iOS 7+ style receipts for auto-renewable subscriptions, too, except as
an array instead of a single receipt object. Given the lack of clear
documentation for this property, I would recommend against using it.